Fin semaine 03

+++ S03TipsCalculator/app/src/main/java/ca/qc/cstj/tipscalculator/domain/models/TipsCalculation.kt
@@ -4,7 +4,7 @@ import ca.qc.cstj.tipscalculator.core.Constants
 
 data class TipsCalculation(private val subtotal: Double, private val tipsPercent: Int)  {
 
-    val tipsAmount = subtotal * (tipsPercent / Constants.PERCENT)
+    val tipsAmount = subtotal * (tipsPercent.toFloat() / Constants.PERCENT)
     val TPS = subtotal * Constants.TPS
     val TVQ = subtotal * Constants.TVQ
     val total = subtotal + tipsAmount + TPS + TVQ
